move upload logic to model
